# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fighters API",
    "description": "API for managing fighters and users with JWT authentication",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "User unique identifier"
          },
          "name": {
            "type": "string",
            "description": "User's full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "cpf": {
            "type": "string",
            "description": "User's CPF (Brazilian ID)"
          }
        }
      },
      "Fighter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Fighter unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Fighter's full name"
          },
          "image": {
            "type": "string",
            "description": "Fighter's image filename"
          },
          "weightClass": {
            "type": "string",
            "description": "Fighter's weight class"
          },
          "nationality": {
            "type": "string",
            "description": "Fighter's nationality"
          },
          "specialty": {
            "type": "string",
            "description": "Fighter's main fighting style"
          },
          "record": {
            "type": "string",
            "description": "Fighter's record (wins-losses-draws)"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/users/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "description": "Create a new admin user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "cpf"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "cpf": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login user",
        "description": "Authenticate user and get JWT token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "description": "Get a list of all registered users",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/search": {
      "get": {
        "tags": ["Users"],
        "summary": "Search users",
        "description": "Search for users by email or CPF",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "description": "Email to search for"
          },
          {
            "in": "query",
            "name": "cpf",
            "schema": {
              "type": "string"
            },
            "description": "CPF to search for"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": ["Users"],
        "summary": "Update user",
        "description": "Update user information",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "cpf": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "description": "Delete a user from the system",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "400": {
            "description": "Cannot delete last admin"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/fighters": {
      "get": {
        "tags": ["Fighters"],
        "summary": "List all fighters",
        "description": "Get a list of all fighters",
        "responses": {
          "200": {
            "description": "List of fighters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fighter"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Fighters"],
        "summary": "Create fighter",
        "description": "Add a new fighter",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "weightClass", "nationality", "specialty", "record"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  },
                  "weightClass": {
                    "type": "string"
                  },
                  "nationality": {
                    "type": "string"
                  },
                  "specialty": {
                    "type": "string"
                  },
                  "record": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fighter created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fighter"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/api/fighters/search": {
      "get": {
        "tags": ["Fighters"],
        "summary": "Search fighters",
        "description": "Search for fighters by name or weight class",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "description": "Fighter name to search for"
          },
          {
            "in": "query",
            "name": "weightClass",
            "schema": {
              "type": "string"
            },
            "description": "Weight class to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching fighters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fighter"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/fighters/{id}": {
      "put": {
        "tags": ["Fighters"],
        "summary": "Update fighter",
        "description": "Update fighter information",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Fighter ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  },
                  "weightClass": {
                    "type": "string"
                  },
                  "nationality": {
                    "type": "string"
                  },
                  "specialty": {
                    "type": "string"
                  },
                  "record": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fighter updated successfully"
          },
          "404": {
            "description": "Fighter not found"
          }
        }
      },
      "delete": {
        "tags": ["Fighters"],
        "summary": "Delete fighter",
        "description": "Delete a fighter from the system",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Fighter ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Fighter deleted successfully"
          },
          "404": {
            "description": "Fighter not found"
          }
        }
      }
    }
  }
}